requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
dj-database-url==2.1.0
Django==5.0.4
django-crispy-forms==2.1
django-filter==24.2
gunicorn==21.2.0
packaging==24.0
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.11.0
